Reworked new fetch methods.
