+++ src/utils/rest-api/index.js
@@ -1,46 +1,4 @@
-import request from 'utils/request'
-
-/**
- * Returns a route with the provided args.
- *
- * @since 0.1
- * @param {String} route
- * @param {Object} args
- * @return {Object}
- */
-export function getRoute( route, args = {} ) {
-	const keys = Object.keys( args )
-
-	if ( keys.length ) {
-		route += '?'
-	}
-
-	keys.map( ( key, index ) => {
-		route += `${ key }=${ encodeURIComponent( args[ key ] ) }`
-		if ( index < keys.length - 1 ) {
-			route += '&'
-		}
-	} )
-
-	return route
-}
-
-/**
- * Makes a get request for an endpoint with the
- * provided args added automatically.
- *
- * @since 0.1
- * @param {String} endpoint
- * @param {Object} args
- * @param {Function} complete
- * @return {Object}
- */
-export function getEndpoint( endpoint, args, complete ) {
-	return request( {
-		route: getRoute( `fl-assistant/v1/${ endpoint }`, args ),
-		complete
-	} )
-}
+import { request, addQueryArgs } from 'utils/request'
 
 /**
  * Returns any array of content for the given type
@@ -70,7 +28,10 @@ export function getContent( type, args, complete ) {
  * @return {Object}
  */
 export function getPosts( args, complete ) {
-	return getEndpoint( 'posts', args, complete )
+	return request( {
+		route: addQueryArgs( 'fl-assistant/v1/posts', args ),
+		complete
+	} )
 }
 
 /**
@@ -82,7 +43,10 @@ export function getPosts( args, complete ) {
  * @return {Object}
  */
 export function getTerms( args, complete ) {
-	return getEndpoint( 'terms', args, complete )
+	return request( {
+		route: addQueryArgs( 'fl-assistant/v1/terms', args ),
+		complete
+	} )
 }
 
 /**
@@ -94,5 +58,8 @@ export function getTerms( args, complete ) {
  * @return {Object}
  */
 export function getUsers( args, complete ) {
-	return getEndpoint( 'users', args, complete )
+	return request( {
+		route: addQueryArgs( 'fl-assistant/v1/users', args ),
+		complete
+	} )
 }
